yum.exclude usage and documentation is not clear.
    Updated documentation rot to fix several inconsistencies.  Fix,
yum.exclude and yum.installonlypkgs lack of accurate documentation.
Add Array.new to exclude and installonlypkg attributes to make this
intention more clear.  Update metadata documentation to accurately
reflect these are Arrays, and not Strings.  Update README to reflect
these are Arrays, and their default value is empty.

    The default recipe has not done anything since shortly after this
cookbook's release by opscode.  Update documentation to accurately
reflect what the default recipe does, which is nothing.  Remove all
mention of `yum update` updating the package cache.  The LWRP updates
the package cache, with `yum -q makecache` when a new repo is added.

    Fix, within the templates, change `if <string>` conditional to
`unless <array>.empty?`.  This reads more clear and conveys the join
method which follows is meant to operate on an Array attribute.