[COOK-2946] - update tests, test kitchen support
This commit updates the baseboxes used for Test Kitchen to be
provisionerless, requiring use of the chef-omnibus package for
installation. This is handled in the .kitchen.yml file. It also
updates the versions to the latest released CentOS 5.9 and 6.4.

The "test cookbook" is removed, being replaced by a single "test"
recipe. The multiple suites are replaced with a single test suite. The
functionality of the various repository recipes are isolated to the
individual files that they manage, so they can be tested safely
together.

In testing, it was revealed that the IUS release attribute was
outdated, and needed to be incremented.

A new helper is added to check that repositories are enabled according
to yum itself.