COOK-3025 Allow per-repo proxy in yum_repository LWRP
While yum::yum allows setting the main global proxy URL, one
can also set this per-repository according to the yum.conf
man page. This is useful, most commonly when set to _none_,
to avoid hitting a proxy for intranet repositories. This
commit adds 'proxy' as an allowed parameter to the yum_repository
LWRP.