[COOK-4164] Revert yum_key
The 3.0 yum removes the yum_key LWRP on the basis that yum will
automatically import gpgkeys if set to a remote URL. This is only true,
if `assumeyes` (Assume Yes) is set globally in the central yum.conf.

There is no "Assume Yes" for importing GPG keys, or for individual repo
files. Therefore, setting `assumeyes` globally is an, all or nothing,
action that would change default behavior for all other yum commands,
e.g. `yum upgrade`.

Proposed restoration of functionality of yum_key LWRP from commitish
11d24bfea28455e55484ecbce407eebdb92b8adc. Otherwise, the yum_repository
LWRP could be modified to accommodate the shortcomings of yum and the
importation of GPG keys when `assumeyes` is set to false globally.

Rubocop:
* Always use hash rockets in hashes.

Foodcritic:
* FC023: Prefer conditional attributes: ./providers/key.rb