Use file resource to delete repo file
When using template resource to delete repo file you may receive an error that the template source file does not exist.  This is because the source is not being specified so Chef is looking for it in a default location based on the resource name.  Changing the resource type to file addresses the issue.