Dynamically generate the new_resource attributes
Issue #99 came up because an attribute for the `yum_globalconfig` was
left out - the cachedir. This commit updates the recipe to dynamically
walk the `node['yum']['main']` attributes and send each one with its
value to the new resource. This assures that all attributes defined
are passed into the configuration.

This *can* lead to an error if a user adds an attribute incorrectly.
That seems unlikely, though. I hope.