Flush internal package cache after module action
After switching to a new module installed during the Chef run, Chef's internal
package cache won't pick up on new module packages automatically, so we need
to reload that manually, much like after adding a new repo.

This isn't needed for modules available at the start of the Chef run, which is
why this was not needed for the existing tests with all default modules.

Signed-off-by: Robert Detjens <detjensrobert@osuosl.org>